ai_backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.4.0
python-dotenv==1.0.1
google-genai>=0.8.0
google-api-core==2.21.0
aiosqlite==0.20.0
sqlalchemy[asyncio]==2.0.35
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
httpx==0.27.2
groq>=0.4.0